move menu to translation module
