+++ src/packages/dictionary/manifests.ts
@@ -1,7 +1,7 @@
 import { manifests as collectionManifests } from './collection/manifests.js';
 import { manifests as dashboardManifests } from './dashboard/manifests.js';
 import { manifests as entityActionManifests } from './entity-action/manifests.js';
-import { manifests as menuManifests } from './menu/manifests.js';
+import { manifests as menuItemManifests } from './menu-item/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
 import { manifests as treeManifests } from './tree/manifests.js';
 import { manifests as workspaceManifests } from './workspace/manifests.js';
@@ -10,7 +10,7 @@ export const manifests: Array<UmbExtensionManifest> = [
 	...collectionManifests,
 	...dashboardManifests,
 	...entityActionManifests,
-	...menuManifests,
+	...menuItemManifests,
 	...repositoryManifests,
 	...treeManifests,
 	...workspaceManifests,
