+++ src/packages/translation/index.ts
@@ -1,1 +1,2 @@
-export * from './constants.js';
+export * from './section/index.js';
+export * from './menu/index.js';
