+++ src/packages/translation/manifests.ts
@@ -1,20 +1,4 @@
-import { UMB_TRANSLATION_SECTION_ALIAS } from './constants.js';
+import { manifests as sectionManifests } from './section/manifests.js';
+import { manifests as menuManifests } from './menu/manifests.js';
 
-export const manifests: Array<UmbExtensionManifest> = [
-	{
-		type: 'section',
-		alias: UMB_TRANSLATION_SECTION_ALIAS,
-		name: 'Translation Section',
-		weight: 400,
-		meta: {
-			label: '#sections_translation',
-			pathname: 'translation',
-		},
-		conditions: [
-			{
-				alias: 'Umb.Condition.SectionUserPermission',
-				match: UMB_TRANSLATION_SECTION_ALIAS,
-			},
-		],
-	},
-];
+export const manifests: Array<UmbExtensionManifest> = [...sectionManifests, ...menuManifests];
