+++ src/packages/translation/menu/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_TRANSLATION_MENU_ALIAS = 'Umb.Menu.Translation';
