+++ src/packages/translation/menu/manifests.ts
@@ -1,0 +1,9 @@
+import { UMB_TRANSLATION_MENU_ALIAS } from './constants.js';
+
+export const manifests: Array<UmbExtensionManifest> = [
+	{
+		type: 'menu',
+		alias: UMB_TRANSLATION_MENU_ALIAS,
+		name: 'Translation Menu',
+	},
+];
